requirements.txt
google-genai
python-dotenv
discord.py
pandas